Ingenu

+++ Ingenu_Power/Ingenu_Power/Domain/StaticInfor.cs
@@ -170,7 +170,7 @@ namespace Ingenu_Power.Domain
 			/// <summary>
 			/// 主电欠压点，用于全项测试
 			/// </summary>
-			public int Voltage_SourceChange_MpUnderVoltage;
+			public decimal Voltage_SourceChange_MpUnderVoltage;
 			/// <summary>
 			/// 主电欠压恢复从备电切换到主电功能合格检查
 			/// </summary>
@@ -178,7 +178,7 @@ namespace Ingenu_Power.Domain
 			/// <summary>
 			/// 主电欠压恢复点，用于全项测试
 			/// </summary>
-			public int Voltage_SourceChange_MpUnderVoltageRecovery;
+			public decimal Voltage_SourceChange_MpUnderVoltageRecovery;
 			/// <summary>
 			/// 主电过压切换到备电功能合格检查
 			/// </summary>
@@ -186,7 +186,7 @@ namespace Ingenu_Power.Domain
 			/// <summary>
 			/// 主电过压点，用于全项测试
 			/// </summary>
-			public int Voltage_SourceChange_MpOverVoltage;
+			public decimal Voltage_SourceChange_MpOverVoltage;
 			/// <summary>
 			/// 主电过压恢复从备电切换到主电功能合格检查
 			/// </summary>
@@ -194,7 +194,7 @@ namespace Ingenu_Power.Domain
 			/// <summary>
 			/// 主电过压恢复点，用于全项测试
 			/// </summary>
-			public int Voltage_SourceChange_MpOverVoltageRecovery;
+			public decimal Voltage_SourceChange_MpOverVoltageRecovery;
 		};
